+++ tests/model.rs
@@ -9,6 +9,13 @@ fn test_design_web_format_xps() {
 
 #[cfg(feature = "reader-xml")]
 #[test]
+fn test_digital_asset_exchange() {
+    let format = FileFormat::from_file("fixtures/model/sample.dae").unwrap();
+    assert_eq!(format, FileFormat::DigitalAssetExchange);
+}
+
+#[cfg(feature = "reader-xml")]
+#[test]
 fn test_extensible_3d_graphics() {
     let format = FileFormat::from_file("fixtures/model/sample.x3d").unwrap();
     assert_eq!(format, FileFormat::Extensible3DGraphics);